variadic argument and expand index support with map.
map<I>(function, source1, ..., sourceI, ..., sourceN)
will expand the elements in sourceI as shown in the following pseudo-code:
valueI = value-of(sourceI)
output = {}
foreach element in valueI
  result = function(value-of(source1), ..., value-of(sourceI-1),
    element, value-of(sourceI+1), ..., value-of(sourceN))
  append(output, result)

Note that, as always, each invocation of function is done independently
and lock-free on the executor.

filter and map_cache do not yet have support for multiple dependencies.