variadic argument and expand index support with filter.
filter<I>(function, sourrce1, ..., sourceI, ..., sourceN)
will expand the elements in sourceI as shown in the following pseudo-code:
valueI = value-of(sourceI)
output = {}
foreach element in valueI
  if (function(value-of(source1), ..., value-of(sourceI-1),
      element, value-of(sourceI+1), ..., value-of(sourceN)))
    append(output, element)

Note that sources not including sourceI are only used in invocation of
function. Not in the output.

Note that, as always, each invocation of function is done independently
and lock-free on the executor. There is no ordering guarantee on the
output collection.