variadic argument and expand index support with map_cache. map_cache<I>(function, source1, ..., sourceI, ..., sourceN) similar to map, but reuses container values from previous iterations if source1, ..., sourceI-1, sourceI+1, ..., sourceN has not changed and container elements in sourceI are equal to the previous.
In other words, if sourceJ where J != I has not changed (revision is same)
and the element of the container value of sourceI is the same as previously,
the call to function is suppressed and the resulting value from the previous
invocation is used.

Note that, as always, each invocation of function is done independently
and lock-free on the executor.